Reference

Index

TwoBandHamiltonians.GappedDirac

TwoBandHamiltonians.GeneralTwoBand

TwoBandHamiltonians.UnitScaling

TwoBandHamiltonians.dx_cc

TwoBandHamiltonians.dx_cv

TwoBandHamiltonians.dx_vc

TwoBandHamiltonians.dx_vv

TwoBandHamiltonians.dy_cc

TwoBandHamiltonians.dy_cv

TwoBandHamiltonians.dy_vc

TwoBandHamiltonians.dy_vv

TwoBandHamiltonians.hvec

TwoBandHamiltonians.vx_cc

TwoBandHamiltonians.vx_cv

TwoBandHamiltonians.vx_vc

TwoBandHamiltonians.vx_vv

TwoBandHamiltonians.vy_cc

TwoBandHamiltonians.vy_cv

TwoBandHamiltonians.vy_vc

TwoBandHamiltonians.vy_vv

TwoBandHamiltonians.Δϵ

TwoBandHamiltonians.ϵ

Docstrings

TwoBandHamiltonians.GappedDirac — Type

GappedDirac{T<:Real} <: GeneralTwoBand{T}

Holds the parameter of a dimensionless massive Dirac Hamiltonian.

The Hamiltonian reads

\[\hat{H} = k_x\sigma_x + k_y\sigma_y + m\sigma_z\]

such that $\vec{h}=[k_x,k_y,m]$.

Examples

julia> h = GappedDirac(1.0)
GappedDirac{Float64}(1.0)

See also

GeneralTwoBand

TwoBandHamiltonians.GeneralTwoBand — Type

GeneralTwoBand{T} <: Hamiltonian{T}

Supertype of all 2x2 Hamiltonians with all matrixelements via dispatch.

Idea

The idea is that all Hamiltonians of the form

\[\hat{H} = \vec{h}(\vec{k})\cdot\vec{\sigma}\]

can be diagonalized analytically and hence most desired matrixelements such as velocities or dipoles can be expressed solely through

\[\vec{h}(\vec{k})=[h_x(\vec{k}),h_y(\vec{k}),h_z(\vec{k})]\]

and its derivatives with respect to $k_\mu$. Any particular Hamiltonian deriving form GeneralTwoBand{T} must then only implement $\vec{h}(\vec{k})$ and its derivatives.

See also

GappedDirac

TwoBandHamiltonians.UnitScaling — Type

UnitScaling(timescale,lengthscale)

Represents a physical length- and time-scale used for non-dimensionalization of a system.

Examples

julia> using Unitful; us = UnitScaling(u"1.0s",u"1.0m")
UnitScaling{Float64}(1.0e15, 1.0e9)

Further information

See here

TwoBandHamiltonians.dx_cc — Method

dx_cc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|x|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_cc(h,1.0,-1.0)
-0.10566243270259357

TwoBandHamiltonians.dx_cv — Method

dx_cv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|x|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_cv(h,1.0,-1.0)
-0.0610042339640731 - 0.22767090063073978im

TwoBandHamiltonians.dx_vc — Method

dx_vc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|x|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_vc(h,1.0,-1.0)
-0.0610042339640731 + 0.22767090063073978im

TwoBandHamiltonians.dx_vv — Method

dx_vv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|x|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dx_vv(h,1.0,-1.0)
0.10566243270259357

TwoBandHamiltonians.dy_cc — Method

dy_cc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|y|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_cc(h,1.0,-1.0)
-0.10566243270259357

TwoBandHamiltonians.dy_cv — Method

dy_cv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ+|y|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_cv(h,1.0,-1.0)
-0.22767090063073978 - 0.0610042339640731im

TwoBandHamiltonians.dy_vc — Method

dy_vc(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|y|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_vc(h,1.0,-1.0)
-0.22767090063073978 + 0.0610042339640731im

TwoBandHamiltonians.dy_vv — Method

dy_vv(h,kx,ky)

Returns the dipole operator matrix element ⟨ψ-|y|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); dy_vv(h,1.0,-1.0)
0.10566243270259357

TwoBandHamiltonians.hvec — Method

hvec(h,kx,ky)

Returns $\vec{h}(\vec{k})$ defining the Hamiltonian at $\vec{k}=[k_x,k_y]$.

TwoBandHamiltonians.vx_cc — Method

vx_cc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vx|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_cc(h,1.0,-1.0)
0.5773502691896258

TwoBandHamiltonians.vx_cv — Method

vx_cv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vx|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_cv(h,1.0,-1.0)
0.7886751345948129 - 0.21132486540518708im

TwoBandHamiltonians.vx_vc — Method

vx_vc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vx|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_vc(h,1.0,-1.0)
0.7886751345948129 + 0.21132486540518708im

TwoBandHamiltonians.vx_vv — Method

vx_vv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vx|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vx_vv(h,1.0,-1.0)
-0.5773502691896258

TwoBandHamiltonians.vy_cc — Method

vy_cc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vy|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_cc(h,1.0,-1.0)
-0.5773502691896258

TwoBandHamiltonians.vy_cv — Method

vy_cv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ+|vy|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_cv(h,1.0,-1.0)
0.21132486540518708 - 0.7886751345948129im

TwoBandHamiltonians.vy_vc — Method

vy_vc(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vy|ψ+⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_vc(h,1.0,-1.0)
0.21132486540518708 + 0.7886751345948129im

TwoBandHamiltonians.vy_vv — Method

vy_vv(h,kx,ky)

Returns the velocity operator matrix element ⟨ψ-|vy|ψ-⟩ at $\vec{k}=[k_x,k_y]$.

Example

julia> h = GappedDirac(1.0); vy_vv(h,1.0,-1.0)
0.5773502691896258

TwoBandHamiltonians.Δϵ — Method

Δϵ(h,kx,ky)

Returns the band energy (valence & conduction) difference at $\vec{k}=[k_x,k_y]$.

TwoBandHamiltonians.ϵ — Method

ϵ(h,kx,ky)

Returns the eigenenergy of the positive (conduction band) state at $\vec{k}=[k_x,k_y]$.