Reference

Index

TwoBandHamiltonians.GappedDirac

TwoBandHamiltonians.GappedDiracOld

TwoBandHamiltonians.GeneralTwoBand

TwoBandHamiltonians.UnitScaling

Docstrings

TwoBandHamiltonians.GappedDirac — Type

GappedDirac{T<:Real} <: GeneralTwoBand{T}

Holds the parameter of a dimensionless massive Dirac Hamiltonian.

The Hamiltonian reads

\[\hat{H} = k_x\sigma_x + k_y\sigma_y + m\sigma_z\]

such that $\vec{h}=[k_x,k_y,m]$.

See also

GeneralTwoBand

TwoBandHamiltonians.GappedDiracOld — Type

GappedDiracOld{T<:Real} <: Hamiltonian{T}

Legacy hard-coded. Holds the parameter of a dimensionless massive Dirac Hamiltonian.

The Hamiltonian reads $\hat{H} = k_x\sigma_x + k_y\sigma_y + \Delta\sigma_z$ such that $2\Delta$ is the bandgap at $\vec{k}=0$

Examples

julia> h = GappedDiracOld(0.1)
GappedDiracOld{Float64}(0.1)

See also

UnitScaling(timescale,lengthscale)

TwoBandHamiltonians.GeneralTwoBand — Type

GeneralTwoBand{T} <: Hamiltonian{T}

Supertype of all 2x2 Hamiltonians with all matrixelements via dispatch.

The idea is that all Hamiltonians of the form

\[\hat{H} = \vec{h}(\vec{k})\cdot\vec{\sigma}\]

can be diagonalized analytically and hence most desired matrixelements such as velocities or dipoles can be expressed solely through

\[\vec{h}(\vec{k})=[h_x(\vec{k}),h_y(\vec{k}),h_z(\vec{k})]\]

and its derivatives with respect to $k_\mu$. Any particular Hamiltonian deriving form GeneralTwoBand{T} must then only implement $\vec{h}(\vec{k})$ and its derivatives.

See also

GappedDirac

TwoBandHamiltonians.UnitScaling — Type

UnitScaling(timescale,lengthscale)

Represents a physical length- and time-scale used for non-dimensionalization of a system.

Examples

julia> using Unitful; us = UnitScaling(u"1.0s",u"1.0m")
UnitScaling{Float64}(1.0e15, 1.0e9)

Further information

See here